@@ -2,7 +2,7 @@ module.exports = {
 	pages: {
 		index: {
 			entry: 'src/main.js',
-			title: 'Vue 2'
+			title: 'Vue 3'
 		}
 	}
 }
